requirements.txt
anthropic==0.16.0
python-dotenv==1.0.0
flask==2.3.3
twilio==8.4.0
elevenlabs==0.2.27
pydub==0.25.1
requests==2.31.0
websockets==11.0.3